Refactor - breaking up more.
