+++ css/editor.scss
@@ -1,4 +1,4 @@
-@import 'global/variables';
+// @import 'global/variables';
 @import 'components/post-select-modal';
 
-@import 'node_modules/react-select/dist/react-select.css';
+@import 'node_modules/react-select/dist/react-select';
