+++ js/components/post-select/term-filters.js
@@ -1,0 +1,6 @@
+const termFilters = [
+	{ slug: 'category', rest: 'categories', plural: 'Categories', collection: 'Categories' },
+	{ slug: 'tag', rest: 'tags', plural: 'Tags', collection: 'Tags' },
+];
+
+export default termFilters;
